### 2025-04-28 #8125 @editorialbot
[PRE REVIEW]: PyLiPD: A python package for the manipulation of paleoclimate datasets
**Submitting author:** @khider (<a href="http://orcid.org/0000-0001-7501-8430">Deborah Khider</a>)
**Repository:** https://github.com/LinkedEarth/pylipd
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.0
