### 2025-05-30 #8125 @khider
Do I need to do anything for editor assignment?
