### 2025-06-13 ccee8d0
Added docstrings to generated classes for better readthedocs
