+++ pylipd/classes/chrondata.py
@@ -9,10 +9,18 @@ from pylipd.classes.datatable import DataTable
 from pylipd.classes.model import Model
 
 class ChronData:
+    """PyLiPD representation of the `ChronData` concept from the LinkedEarth ontology.
+
+    This class is AUTO-GENERATED by `create_classes.py`; do not edit by hand.
+    Use the public getter/setter helpers to manipulate its fields.
+    """
 
     def __init__(self):
-        self.measurementTables: list[DataTable] = []
-        self.modeledBy: list[Model] = []
+        """Instantiate a blank \`ChronData\` object with a freshly minted identifier.
+
+        All ontology-defined properties are initialised to `None` (for single-valued
+        properties) or an empty list (for multi-valued properties).
+        """
         self.misc = {}
         self.ontns = "http://linked.earth/ontology#"
         self.ns = "http://linked.earth/lipd"
@@ -21,6 +29,21 @@ class ChronData:
 
     @staticmethod
     def from_data(id, data) -> 'ChronData':
+        """Create a `ChronData` instance from a JSON-LD data dictionary produced by
+        :py:meth:`to_data`.
+
+        Parameters
+        ----------
+        id : str
+            The URI of the node representing this object in the graph.
+        data : dict
+            Complete JSON-LD graph keyed by node id.
+
+        Returns
+        -------
+        ChronData
+            A populated object.
+        """
         self = ChronData()
         self.id = id
         mydata = data[id]
@@ -60,6 +83,19 @@ class ChronData:
         return self
 
     def to_data(self, data={}):
+        """Serialise the object to the LiPD JSON-LD graph structure used internally.
+
+        Parameters
+        ----------
+        data : dict, optional
+            Graph dictionary to append the representation into. New dictionaries
+            are created on demand.
+
+        Returns
+        -------
+        dict
+            Updated graph dictionary.
+        """
         data[self.id] = {}
         data[self.id]["type"] = [
             {
@@ -129,6 +165,9 @@ class ChronData:
         return data
 
     def to_json(self):
+        """Return a lightweight, user-friendly JSON representation of the object.
+        Intended for external consumption (e.g., saving to file or UI rendering).
+        """
         data = {
             "@id": self.id
         }
@@ -153,6 +192,19 @@ class ChronData:
 
     @staticmethod
     def from_json(data) -> 'ChronData':
+        """Re-instantiate a `ChronData` from the lightweight JSON produced by
+        :py:meth:`to_json`.
+
+        Parameters
+        ----------
+        data : dict
+            JSON representation.
+
+        Returns
+        -------
+        ChronData
+            Populated instance.
+        """
         self = ChronData()
         for key in data:
             pvalue = data[key]
@@ -172,41 +224,99 @@ class ChronData:
         return self
 
     def set_non_standard_property(self, key, value):
+        """Assign a non-ontology property.
+
+        These *misc* properties are preserved during serialisation but have no
+        dedicated first-class attribute on the object.
+        """
         if key not in self.misc:
             self.misc[key] = value
     
     def get_non_standard_property(self, key):
+        """Retrieve a previously stored non-ontology property.
+
+        Raises
+        ------
+        KeyError
+            If the property is not present.
+        """
         return self.misc[key]
                 
     def get_all_non_standard_properties(self):
+        """Return the entire miscellaneous property dictionary."""
         return self.misc
 
     def add_non_standard_property(self, key, value):
+        """Append *value* to a list-valued miscellaneous property identified by
+        *key*. Creates the list when needed.
+        """
         if key not in self.misc:
             self.misc[key] = []
         self.misc[key].append(value)
         
     def getMeasurementTables(self) -> list[DataTable]:
[…]
