+++ pylipd/classes/archivetype.py
@@ -5,6 +5,9 @@
 from pylipd.globals.synonyms import SYNONYMS
 
 class ArchiveType:
+    """Enumeration helper representing LiPD controlled vocabulary term group `ArchiveType`.
+    AUTO-GENERATED – do not modify by hand.
+    """
     synonyms = SYNONYMS["ARCHIVES"]["ArchiveType"]
 
     def __init__(self, id, label):
@@ -15,12 +18,15 @@ class ArchiveType:
             self.id == value.id
         
     def getLabel(self):
+        """Return the human-readable label for this enumeration value."""
         return self.label
 
     def getId(self):
+        """Return the identifier/URI for this enumeration value."""
         return self.id
     
     def to_data(self, data={}):
+        """Serialise this enumeration value to the internal JSON-LD graph format."""
         data[self.id] ={
             "label": [
                 {
@@ -33,17 +39,22 @@ class ArchiveType:
         return data
 
     def to_json(self):
+        """Return a minimal JSON value (string) corresponding to this synonym."""
         data = self.label
         return data
 
     @classmethod
     def from_synonym(cls, synonym):
+        """Return a new `ArchiveType` instance matching a synonym string, or `None`."""
         if synonym.lower() in ArchiveType.synonyms:
             synobj = ArchiveType.synonyms[synonym.lower()]
             return ArchiveType(synobj['id'], synobj['label'])
         return None
         
 class ArchiveTypeConstants:
+    """Namespace-style container holding pre-instantiated ArchiveType enumeration values.
+    Each attribute corresponds to one controlled vocabulary entry.
+    """
     Borehole = ArchiveType("http://linked.earth/ontology/archive#Borehole", "Borehole")
     Coral = ArchiveType("http://linked.earth/ontology/archive#Coral", "Coral")
     FluvialSediment = ArchiveType("http://linked.earth/ontology/archive#FluvialSediment", "Fluvial sediment")
