+++ pylipd/classes/person.py
@@ -7,9 +7,18 @@ import re
 from pylipd.utils import uniqid
 
 class Person:
+    """PyLiPD representation of the `Person` concept from the LinkedEarth ontology.
+
+    This class is AUTO-GENERATED by `create_classes.py`; do not edit by hand.
+    Use the public getter/setter helpers to manipulate its fields.
+    """
 
     def __init__(self):
-        self.name: str = None
+        """Instantiate a blank \`Person\` object with a freshly minted identifier.
+
+        All ontology-defined properties are initialised to `None` (for single-valued
+        properties) or an empty list (for multi-valued properties).
+        """
         self.misc = {}
         self.ontns = "http://linked.earth/ontology#"
         self.ns = "http://linked.earth/lipd"
@@ -18,6 +27,21 @@ class Person:
 
     @staticmethod
     def from_data(id, data) -> 'Person':
+        """Create a `Person` instance from a JSON-LD data dictionary produced by
+        :py:meth:`to_data`.
+
+        Parameters
+        ----------
+        id : str
+            The URI of the node representing this object in the graph.
+        data : dict
+            Complete JSON-LD graph keyed by node id.
+
+        Returns
+        -------
+        Person
+            A populated object.
+        """
         self = Person()
         self.id = id
         mydata = data[id]
@@ -45,6 +69,19 @@ class Person:
         return self
 
     def to_data(self, data={}):
+        """Serialise the object to the LiPD JSON-LD graph structure used internally.
+
+        Parameters
+        ----------
+        data : dict, optional
+            Graph dictionary to append the representation into. New dictionaries
+            are created on demand.
+
+        Returns
+        -------
+        dict
+            Updated graph dictionary.
+        """
         data[self.id] = {}
         data[self.id]["type"] = [
             {
@@ -90,6 +127,9 @@ class Person:
         return data
 
     def to_json(self):
+        """Return a lightweight, user-friendly JSON representation of the object.
+        Intended for external consumption (e.g., saving to file or UI rendering).
+        """
         data = {
             "@id": self.id
         }
@@ -107,6 +147,19 @@ class Person:
 
     @staticmethod
     def from_json(data) -> 'Person':
+        """Re-instantiate a `Person` from the lightweight JSON produced by
+        :py:meth:`to_json`.
+
+        Parameters
+        ----------
+        data : dict
+            JSON representation.
+
+        Returns
+        -------
+        Person
+            Populated instance.
+        """
         self = Person()
         for key in data:
             pvalue = data[key]
@@ -122,24 +175,54 @@ class Person:
         return self
 
     def set_non_standard_property(self, key, value):
+        """Assign a non-ontology property.
+
+        These *misc* properties are preserved during serialisation but have no
+        dedicated first-class attribute on the object.
+        """
         if key not in self.misc:
             self.misc[key] = value
     
     def get_non_standard_property(self, key):
+        """Retrieve a previously stored non-ontology property.
+
+        Raises
+        ------
+        KeyError
+            If the property is not present.
+        """
         return self.misc[key]
                 
     def get_all_non_standard_properties(self):
+        """Return the entire miscellaneous property dictionary."""
         return self.misc
 
     def add_non_standard_property(self, key, value):
+        """Append *value* to a list-valued miscellaneous property identified by
+        *key*. Creates the list when needed.
+        """
         if key not in self.misc:
             self.misc[key] = []
         self.misc[key].append(value)
         
     def getName(self) -> str:
+        """Get name.
+
+        Returns
+        -------
+        str
+            The current value of name.
+        """
         return self.name
 
     def setName(self, name:str):
+        """Set name.
+
+        Parameters
+        ----------
+        name : str
+            The value to assign.
+        """
         assert isinstance(name, str), f"Error: '{name}' is not of type str"
         self.name = name
     

[…]
