+++ pylipd/classes/datatable.py
@@ -11,18 +11,13 @@ from pylipd.utils import uniqid
 from pylipd.classes.variable import Variable
 
 class DataTable:
-    """PyLiPD representation of the `DataTable` concept from the LinkedEarth ontology.
-
-    This class is AUTO-GENERATED by `create_classes.py`; do not edit by hand.
-    Use the public getter/setter helpers to manipulate its fields.
-    """
-
+    """Auto-generated LinkedEarth class representing `DataTable`."""
     def __init__(self):
-        """Instantiate a blank \`DataTable\` object with a freshly minted identifier.
-
-        All ontology-defined properties are initialised to `None` (for single-valued
-        properties) or an empty list (for multi-valued properties).
-        """
+        """Initialize a new DataTable instance."""
+        
+        self.fileName: str = None
+        self.missingValue: str = None
+        self.variables: list[Variable] = []
         self.misc = {}
         self.ontns = "http://linked.earth/ontology#"
         self.ns = "http://linked.earth/lipd"
@@ -31,20 +26,19 @@ class DataTable:
 
     @staticmethod
     def from_data(id, data) -> 'DataTable':
-        """Create a `DataTable` instance from a JSON-LD data dictionary produced by
-        :py:meth:`to_data`.
+        """Instantiate `DataTable` from an ontology-style data graph.
 
         Parameters
         ----------
         id : str
-            The URI of the node representing this object in the graph.
+            The node identifier for this object.
         data : dict
-            Complete JSON-LD graph keyed by node id.
+            Dictionary mapping node ids to their predicate lists.
 
         Returns
         -------
         DataTable
-            A populated object.
+            The populated `DataTable` instance.
         """
         self = DataTable()
         self.id = id
@@ -88,18 +82,17 @@ class DataTable:
         return self
 
     def to_data(self, data={}):
-        """Serialise the object to the LiPD JSON-LD graph structure used internally.
+        """Serialize the object into a JSON-LD compatible dictionary.
 
         Parameters
         ----------
         data : dict, optional
-            Graph dictionary to append the representation into. New dictionaries
-            are created on demand.
+            Existing data dictionary to extend.
 
         Returns
         -------
         dict
-            Updated graph dictionary.
+            The updated data dictionary.
         """
         data[self.id] = {}
         data[self.id]["type"] = [
@@ -173,8 +166,12 @@ class DataTable:
         return data
 
     def to_json(self):
-        """Return a lightweight, user-friendly JSON representation of the object.
-        Intended for external consumption (e.g., saving to file or UI rendering).
+        """Return a lightweight JSON representation (used by LipdVerse).
+
+        Returns
+        -------
+        dict
+            A dictionary representation of this object.
         """
         data = {
             "@id": self.id
@@ -204,18 +201,17 @@ class DataTable:
 
     @staticmethod
     def from_json(data) -> 'DataTable':
-        """Re-instantiate a `DataTable` from the lightweight JSON produced by
-        :py:meth:`to_json`.
+        """Instantiate `DataTable` from its lightweight JSON representation.
 
         Parameters
         ----------
         data : dict
-            JSON representation.
+            The JSON dictionary to parse.
 
         Returns
         -------
         DataTable
-            Populated instance.
+            The populated `DataTable` instance.
         """
         self = DataTable()
         for key in data:
@@ -240,31 +236,55 @@ class DataTable:
         return self
 
     def set_non_standard_property(self, key, value):
-        """Assign a non-ontology property.
+        """Store a predicate that is not defined in the ontology schema.
 
-        These *misc* properties are preserved during serialisation but have no
-        dedicated first-class attribute on the object.
+        This is useful for forward-compatibility with new properties that are
+        not yet part of the official schema.
+
+        Parameters
+        ----------
+        key : str
+            The property name.
+        value : any
[…]
