+++ pylipd/classes/location.py
@@ -7,18 +7,25 @@ import re
 from pylipd.utils import uniqid
 
 class Location:
-    """PyLiPD representation of the `Location` concept from the LinkedEarth ontology.
-
-    This class is AUTO-GENERATED by `create_classes.py`; do not edit by hand.
-    Use the public getter/setter helpers to manipulate its fields.
-    """
-
+    """Auto-generated LinkedEarth class representing `Location`."""
     def __init__(self):
-        """Instantiate a blank \`Location\` object with a freshly minted identifier.
-
-        All ontology-defined properties are initialised to `None` (for single-valued
-        properties) or an empty list (for multi-valued properties).
-        """
+        """Initialize a new Location instance."""
+        
+        self.continent: str = None
+        self.coordinates: str = None
+        self.coordinatesFor: None = None
+        self.country: str = None
+        self.countryOcean: str = None
+        self.description: str = None
+        self.elevation: str = None
+        self.geometryType: str = None
+        self.latitude: str = None
+        self.locationName: str = None
+        self.locationType: str = None
+        self.longitude: str = None
+        self.notes: str = None
+        self.ocean: str = None
+        self.siteName: str = None
         self.misc = {}
         self.ontns = "http://linked.earth/ontology#"
         self.ns = "http://linked.earth/lipd"
@@ -27,20 +34,19 @@ class Location:
 
     @staticmethod
     def from_data(id, data) -> 'Location':
-        """Create a `Location` instance from a JSON-LD data dictionary produced by
-        :py:meth:`to_data`.
+        """Instantiate `Location` from an ontology-style data graph.
 
         Parameters
         ----------
         id : str
-            The URI of the node representing this object in the graph.
+            The node identifier for this object.
         data : dict
-            Complete JSON-LD graph keyed by node id.
+            Dictionary mapping node ids to their predicate lists.
 
         Returns
         -------
         Location
-            A populated object.
+            The populated `Location` instance.
         """
         self = Location()
         self.id = id
@@ -152,18 +158,17 @@ class Location:
         return self
 
     def to_data(self, data={}):
-        """Serialise the object to the LiPD JSON-LD graph structure used internally.
+        """Serialize the object into a JSON-LD compatible dictionary.
 
         Parameters
         ----------
         data : dict, optional
-            Graph dictionary to append the representation into. New dictionaries
-            are created on demand.
+            Existing data dictionary to extend.
 
         Returns
         -------
         dict
-            Updated graph dictionary.
+            The updated data dictionary.
         """
         data[self.id] = {}
         data[self.id]["type"] = [
@@ -349,8 +354,12 @@ class Location:
         return data
 
     def to_json(self):
-        """Return a lightweight, user-friendly JSON representation of the object.
-        Intended for external consumption (e.g., saving to file or UI rendering).
+        """Return a lightweight JSON representation (used by LipdVerse).
+
+        Returns
+        -------
+        dict
+            A dictionary representation of this object.
         """
         data = {
             "@id": self.id
@@ -439,18 +448,17 @@ class Location:
 
     @staticmethod
     def from_json(data) -> 'Location':
-        """Re-instantiate a `Location` from the lightweight JSON produced by
-        :py:meth:`to_json`.
+        """Instantiate `Location` from its lightweight JSON representation.
 
         Parameters
         ----------
         data : dict
-            JSON representation.
+            The JSON dictionary to parse.
 
         Returns
         -------
         Location
-            Populated instance.
+            The populated `Location` instance.
         """
         self = Location()
         for key in data:
@@ -523,31 +531,55 @@ class Location:
         return self
 
     def set_non_standard_property(self, key, value):
-        """Assign a non-ontology property.
+        """Store a predicate that is not defined in the ontology schema.
 
-        These *misc* properties are preserved during serialisation but have no
[…]

[…]
