### 2025-06-13 60c7d00
Bump version to 1.4.1 - Add function signature docstrings for ReadTheDocs
