+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pylipd"
-version = "1.4.0"
+version = "1.4.1"
 description = "Python utilities for handling LiPD data"
 readme = "README.md"
 license = { text = "Apache-2.0" }
