+++ CITATION.cff
@@ -8,7 +8,7 @@ authors:
     given-names: "Deborah"
     orcid: "https://orcid.org/0000-0001-7501-8430"
 title: "PyLiPD: a Python package for the manipulation of LiPD datasets"
-version: v1.4.0
+version: v1.4.1
 doi: 10.5281/zenodo.7951201
-date-released: 2025-03-24
+date-released: 2025-01-27
 url: "https://github.com/LinkedEarth/pylipd"
