### 2025-06-20 #8125 @willgearty
:wave: @PennyHow, @eviatarbach, @b8raoult, @spencerkclark, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
