### 2025-06-24 ca7acec
Added script to automatically download and create .nq file from LiPDVerse. Still manual intervention required to upload to GraphDB for now
