+++ download_and_convert_lipdverse.py
@@ -1,0 +1,420 @@
+#!/usr/bin/env python3
+"""
+Download all LiPD files from LiPDverse and convert to RDF (.nq format)
+
+This script:
+1. Downloads the latest LiPD files from https://lipdverse.org/lipdverse/current_version/projectSidebar.html
+2. Extracts them to a local directory
+3. Converts all LiPD files to a single RDF .nq file using PyLiPD
+
+Usage:
+    python download_and_convert_lipdverse.py [--output-dir OUTPUT_DIR] [--nq-filename NQ_FILENAME]
+"""
+
+import os
+import sys
+import argparse
+import requests
+import zipfile
+import tempfile
+from pathlib import Path
+from urllib.parse import urljoin, urlparse
+import re
+from bs4 import BeautifulSoup
+
+from pylipd.lipd import LiPD
+
+
+def find_download_link(base_url="https://lipdverse.org/lipdverse/current_version/projectSidebar.html"):
+    """
+    Find the download link for "Download all LiPD files" from the LiPDverse page.
+    
+    Parameters
+    ----------
+    base_url : str
+        The base URL of the LiPDverse current version page
+        
+    Returns
+    -------
+    str
+        The full download URL for the LiPD files
+    """
+    print(f"Fetching download link from: {base_url}")
+    
+    try:
+        response = requests.get(base_url, timeout=30)
+        response.raise_for_status()
+        
+        soup = BeautifulSoup(response.content, 'html.parser')
+        
+        # Look for links containing "Download all LiPD files" or similar text
+        download_link = None
+        for link in soup.find_all('a', href=True):
+            link_text = link.get_text(strip=True).lower()
+            if 'download all lipd files' in link_text or 'download all' in link_text:
+                href = link['href']
+                # Make sure it's a full URL
+                if href.startswith('http'):
+                    download_link = href
+                else:
+                    download_link = urljoin(base_url, href)
+                break
+        
+        if not download_link:
+            # Fallback: look for .zip files
+            for link in soup.find_all('a', href=True):
+                href = link['href']
+                if href.endswith('.zip') and ('lipd' in href.lower() or 'all' in href.lower()):
+                    if href.startswith('http'):
+                        download_link = href
+                    else:
+                        download_link = urljoin(base_url, href)
+                    break
+        
+        if download_link:
+            print(f"Found download link: {download_link}")
+            return download_link
+        else:
+            raise ValueError("Could not find download link for LiPD files on the page")
+            
+    except Exception as e:
+        print(f"Error fetching download link: {e}")
+        raise
+
+
+def download_file(url, local_filename):
+    """
+    Download a file from URL to local filename with progress indication.
+    
+    Parameters
+    ----------
+    url : str
+        The URL to download from
+    local_filename : str
+        The local file path to save to
+    """
+    print(f"Downloading: {url}")
+    print(f"Saving to: {local_filename}")
+    
+    with requests.get(url, stream=True) as r:
+        r.raise_for_status()
+        total_size = int(r.headers.get('content-length', 0))
+        
+        with open(local_filename, 'wb') as f:
+            downloaded = 0
+            for chunk in r.iter_content(chunk_size=8192):
+                if chunk:
+                    f.write(chunk)
+                    downloaded += len(chunk)
+                    if total_size > 0:
+                        percent = (downloaded / total_size) * 100
[…]
