### 2025-06-24 9aed617
Fixed queries for timeseries essentials
