+++ pylipd/usage_parallel.py
@@ -5,15 +5,6 @@ if __name__ == "__main__":
 
     L = LiPD()
 
-    # L.load_from_dir("examples/data/Euro2k", parallel=True, cutoff=1000)
-    # Lfiltered = L.filter_by_geo_bbox(0,25,50,50)
-    # print(Lfiltered.get_all_dataset_names())
-
-    # Lfiltered = L.filter_by_archive_type("marine")
-    # print(len(Lfiltered.get_all_dataset_names()))
-    # print(Lfiltered.get_all_archiveTypes())
-    # print(Lfiltered.get_all_dataset_names())
-
     # Convert LiPD files to RDF    
     L.convert_lipd_dir_to_rdf(
         local_lipd_dir,
