### 2025-06-27 ea6208f
Release v1.4.2
- Add comprehensive docstrings to auto-generated classes for ReadTheDocs documentation
- Add compilation analysis script for LiPDVerse datasets
- Fix package discovery issues in pyproject.toml
- Enhance support for multiple compilations per dataset
- Add dataset comparison CSV generation functionality
