+++ pylipd/classes/compilation.py
@@ -135,7 +135,7 @@ class Compilation:
         return data
 
     def to_json(self):
-        """Return a lightweight JSON representation (used by LipdVerse).
+        """Return a lightweight JSON representation (used by LiPD).
 
         Returns
         -------
