+++ pylipd/classes/interpretation.py
@@ -369,7 +369,7 @@ class Interpretation:
         return data
 
     def to_json(self):
-        """Return a lightweight JSON representation (used by LipdVerse).
+        """Return a lightweight JSON representation (used by LiPD).
 
         Returns
         -------
