+++ pylipd/classes/calibration.py
@@ -339,7 +339,7 @@ class Calibration:
         return data
 
     def to_json(self):
-        """Return a lightweight JSON representation (used by LipdVerse).
+        """Return a lightweight JSON representation (used by LiPD).
 
         Returns
         -------
