+++ pylipd/classes/funding.py
@@ -180,7 +180,7 @@ class Funding:
         return data
 
     def to_json(self):
-        """Return a lightweight JSON representation (used by LipdVerse).
+        """Return a lightweight JSON representation (used by LiPD).
 
         Returns
         -------
