+++ pylipd/classes/changelog.py
@@ -214,7 +214,7 @@ class ChangeLog:
         return data
 
     def to_json(self):
-        """Return a lightweight JSON representation (used by LipdVerse).
+        """Return a lightweight JSON representation (used by LiPD).
 
         Returns
         -------
