+++ pylipd/classes/resolution.py
@@ -194,7 +194,7 @@ class Resolution:
         return data
 
     def to_json(self):
-        """Return a lightweight JSON representation (used by LipdVerse).
+        """Return a lightweight JSON representation (used by LiPD).
 
         Returns
         -------
