+++ pylipd/classes/dataset.py
@@ -546,7 +546,7 @@ class Dataset:
         return data
 
     def to_json(self):
-        """Return a lightweight JSON representation (used by LipdVerse).
+        """Return a lightweight JSON representation (used by LiPD).
 
         Returns
         -------
