+++ pylipd/classes/physicalsample.py
@@ -152,7 +152,7 @@ class PhysicalSample:
         return data
 
     def to_json(self):
-        """Return a lightweight JSON representation (used by LipdVerse).
+        """Return a lightweight JSON representation (used by LiPD).
 
         Returns
         -------
