+++ pylipd/classes/datatable.py
@@ -166,7 +166,7 @@ class DataTable:
         return data
 
     def to_json(self):
-        """Return a lightweight JSON representation (used by LipdVerse).
+        """Return a lightweight JSON representation (used by LiPD).
 
         Returns
         -------
