### 2025-06-27 206959c
Update CITATION.cff
