### 2025-06-29 cf7c495
New scripts for lipdverse <-> graphdb conversion
