+++ paper.md
@@ -1,49 +1,0 @@
----
-title: 'PyLiPD: A python package for the manipulation of paleoclimate datasets'
-tags:
-  - Python
-  - paleoclimate
-  - knowledge graph
-authors:
-  - name: Varun Ratnakar
-    orcid: 0000-0001-9381-705X
-    affiliation: 1
-  - name: Deborah Khider
-    orcid: 0000-0001-7501-8430
-    affiliation: 1
-affiliations:
- - name: Information Sciences Institute, University of Southern California
-   index: 1
-date: 25 April 2025
-bibliography: paper.bib
----
-
-# Summary
-`PyLiPD` is a Python package designed to support the reading, querying, editing, and writing of LiPD-formatted paleoclimate datasets. Built on the `rdflib` library, it transforms LiPD data into RDF graphs aligned with the LinkedEarth ontology, enabling semantic querying with SPARQL. The API is structured into four modules that handle data access, variable-level filtering, and dataset creation and editing, with support for both graph-based and tabular workflows. Comprehensive documentation and tutorials—including ontology concepts and scientific use cases—are provided to guide users. 
-
-# Statement of Need
-Paleoclimate data (obtained from biogeophysical measurements made on ice cores, sediments, corals, trees, etc.) often come with diverse structures and metadata conventions, making synthesis efforts challenging and highlighting the necessity of standardized data formats.  Achieving standardization involves three key components: (1) a uniform data format, (2) a consistent terminology for describing metadata, and (3) clear guidelines for how data should be reported. The Linked PaleoData (LiPD) framework emerged to address the first need, providing a universally readable data container for paleoclimate data and metadata [@McKay2016]. The metadata is stored in a JSON-LD file while the data is organized in multiple tables saved as csv files. LiPD has six distinct components: root metadata (e.g., dataset name, and version); geographic metadata (e.g., coordinates); publication metadata (including unique identifiers); funding metadata (e.g., grant number); PaleoData, which includes all the measured (e.g., the width of tree rings) and inferred (e.g., temperature) paleoenvironmental data; and ChronData, which mirrors PaleoData for information pertaining to time. These components provide the rigidity necessary to write robust codes around the format while remaining extensible enough to capture (meta)data as rich as the users want to provide for them. Paired with the Paleoenvironmental Standard Terms (PaST) Thesaurus developed by the National Oceanic and Atmospheric Administration (NOAA) [@Morrill2021], LiPD standardization facilitates efficient querying and analysis of multiple records, enabling large-scale paleoclimate syntheses. Community-driven initiatives to reconstruct temperature and hydroclimate over Earth's history [@pages2k; @Kaufman2020; @Routson2021; @Konecky2020, @Jonkers2020] have compiled paleoclimate records into this standardized format and have made them available to the paleoclimate community in an online database. As the database continued to grow, there was an increasing demand for tools capable of accessing, reading, modifying, and writing files in the LiPD format. To address this need, we introduce `PyLiPD`, a Python-based tool designed specifically for these tasks.
-
-# Implementation
-PyLiPD is built on top of the Python `rdflib` library [@Krech_RDFLib_2025], which supports working with Resource Description Framework (RDF) data and provides robust capabilities for parsing, manipulating, and querying semantic graphs using SPARQL. To leverage this functionality, LiPD-formatted datasets are converted into RDF graphs upon loading, using the LinkedEarth ontology [@LinkedEarthOntology]. The LinkedEarth Ontology captures the LiPD components and the relationship among them and integrates the standardized terms from the PaST Thesaurus.  
-
[…]
