+++ paper.bib
@@ -1,147 +1,0 @@
-@article{McKay2016,
-  title = {Technical note: The Linked Paleo Data framework – a common tongue for paleoclimatology},
-  volume = {12},
-  ISSN = {1814-9332},
-  url = {http://dx.doi.org/10.5194/cp-12-1093-2016},
-  DOI = {10.5194/cp-12-1093-2016},
-  number = {4},
-  journal = {Climate of the Past},
-  publisher = {Copernicus GmbH},
-  author = {McKay,  Nicholas P. and Emile-Geay,  Julien},
-  year = {2016},
-  month = apr,
-  pages = {1093–1100}
-}
-
-@article{Khider2019,
-  title = {PaCTS 1.0: A Crowdsourced Reporting Standard for Paleoclimate Data},
-  volume = {34},
-  ISSN = {2572-4525},
-  url = {http://dx.doi.org/10.1029/2019PA003632},
-  DOI = {10.1029/2019pa003632},
-  number = {10},
-  journal = {Paleoceanography and Paleoclimatology},
-  publisher = {American Geophysical Union (AGU)},
-  author = {Khider,  D. and Emile‐Geay,  J. and McKay,  N. P. and Gil,  Y. and Garijo,  D. and Ratnakar,  V. and Alonso‐Garcia,  M. and Bertrand,  S. and Bothe,  O. and Brewer,  P. and Bunn,  A. and Chevalier,  M. and Comas‐Bru,  L. and Csank,  A. and Dassié,  E. and DeLong,  K. and Felis,  T. and Francus,  P. and Frappier,  A. and Gray,  W. and Goring,  S. and Jonkers,  L. and Kahle,  M. and Kaufman,  D. and Kehrwald,  N. M. and Martrat,  B. and McGregor,  H. and Richey,  J. and Schmittner,  A. and Scroxton,  N. and Sutherland,  E. and Thirumalai,  K. and Allen,  K. and Arnaud,  F. and Axford,  Y. and Barrows,  T. and Bazin,  L. and Pilaar Birch,  S. E. and Bradley,  E. and Bregy,  J. and Capron,  E. and Cartapanis,  O. and Chiang,  H.‐W. and Cobb,  K. M. and Debret,  M. and Dommain,  R. and Du,  J. and Dyez,  K. and Emerick,  S. and Erb,  M. P. and Falster,  G. and Finsinger,  W. and Fortier,  D. and Gauthier,  Nicolas and George,  S. and Grimm,  E. and Hertzberg,  J. and Hibbert,  F. and Hillman,  A. and Hobbs,  W. and Huber,  M. and Hughes,  A. L. C. and Jaccard,  S. and Ruan,  J. and Kienast,  M. and Konecky,  B. and Le Roux,  G. and Lyubchich,  V. and Novello,  V. F. and Olaka,  L. and Partin,  J. W. and Pearce,  C. and Phipps,  S. J. and Pignol,  C. and Piotrowska,  N. and Poli,  M.‐S. and Prokopenko,  A. and Schwanck,  F. and Stepanek,  C. and Swann,  G. E. A. and Telford,  R. and Thomas,  E. and Thomas,  Z. and Truebe,  S. and von Gunten,  L. and Waite,  A. and Weitzel,  N. and Wilhelm,  B. and Williams,  J. and Williams,  J. J. and Winstrup,  M. and Zhao,  N. and Zhou,  Y.},
-  year = {2019},
-  month = oct,
-  pages = {1570–1596}
-}
-
-
-@misc{LinkedEarthOntology,
-  doi = {10.5281/ZENODO.2577604},
-  url = {https://zenodo.org/record/2577604},
[…]
