+++ scripts/upload_to_graphdb.py
@@ -4,11 +4,12 @@ Upload data to GraphDB
 
 This script:
 1. Connects to a GraphDB instance
-2. Deletes all existing graphs (except the default graph)
-3. Imports a .nq.zip file with target graphs set to "from data"
+2. Clears the entire repository
+3. Imports all TTL files from ontology/*.ttl to the default graph
+4. Imports a .nq.zip file with target graphs set to "from data"
 
 Usage:
-    python upload_to_graphdb.py --endpoint ENDPOINT --nq-zip-file FILE [--username USER] [--password PASS]
+    python upload_to_graphdb.py --endpoint ENDPOINT --nq-zip-file FILE [--ontology-dir ONTOLOGY_DIR] [--username USER] [--password PASS]
 """
 
 import os
@@ -22,7 +23,7 @@ import json
 
 
 class GraphDBUploader:
-    def __init__(self, endpoint_url, username=None, password=None):
+    def __init__(self, endpoint_url, ontology_dir="ontology", username=None, password=None):
         """
         Initialize GraphDB uploader.
         
@@ -30,6 +31,8 @@ class GraphDBUploader:
         ----------
         endpoint_url : str
             GraphDB repository endpoint URL (e.g., http://localhost:7200/repositories/myrepo)
+        ontology_dir : str
+            Directory containing TTL ontology files
         username : str, optional
             Username for authentication
         password : str, optional
@@ -38,6 +41,7 @@ class GraphDBUploader:
         self.endpoint_url = endpoint_url.rstrip('/')
         self.base_url = '/'.join(endpoint_url.split('/')[:-2])  # Remove /repositories/reponame
         self.repo_name = endpoint_url.split('/')[-1]
+        self.ontology_dir = Path(ontology_dir)
         self.username = username
         self.password = password
         
@@ -48,15 +52,17 @@ class GraphDBUploader:
         
         # Upload statistics
         self.upload_stats = {
-            'graphs_deleted': 0,
-            'import_successful': False,
-            'import_time_seconds': 0,
+            'repository_cleared': False,
+            'ontology_files_imported': 0,
+            'data_import_successful': False,
+            'total_time_seconds': 0,
             'errors': []
         }
         
         print(f"GraphDB Uploader initialized:")
         print(f"  Base URL: {self.base_url}")
         print(f"  Repository: {self.repo_name}")
+        print(f"  Ontology directory: {self.ontology_dir}")
         print(f"  Authentication: {'Yes' if self.auth else 'No'}")
 
     def test_connection(self):
@@ -64,7 +70,7 @@ class GraphDBUploader:
         print("Testing connection to GraphDB...")
         
         try:
-            # Test basic connectivity
+            # Test basic connectivity with REST API
             test_url = f"{self.base_url}/rest/repositories"
             response = requests.get(test_url, auth=self.auth, timeout=10)
             response.raise_for_status()
@@ -76,6 +82,22 @@ class GraphDBUploader:
             if repo_exists:
                 print(f"✓ Connected to GraphDB successfully")
                 print(f"✓ Repository '{self.repo_name}' found")
+                
+                # Test SPARQL endpoint specifically
+                try:
+                    test_query = "SELECT (COUNT(*) as ?count) WHERE { ?s ?p ?o }"
+                    sparql_response = requests.post(
+                        f"{self.endpoint_url}",
+                        data={'query': test_query},
+                        headers={'Accept': 'application/sparql-results+json'},
+                        auth=self.auth,
+                        timeout=10
+                    )
+                    sparql_response.raise_for_status()
+                    print(f"✓ SPARQL endpoint working")
+                except Exception as e:
+                    print(f"⚠ SPARQL endpoint test failed: {e}")
+                
[…]

[…]
