### 2025-06-29 69d4560
Bump version to v1.4.3
- Fix upload_to_graphdb.py to upload .nq.zip files directly without extraction
- Add comprehensive LiPDVerse processing scripts (download, preprocess, convert, upload)
- Enhance GraphDB integration with better error handling and SPARQL operations
- Improve file handling with macOS artifact cleanup
