+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pylipd"
-version = "1.4.2"
+version = "1.4.3"
 description = "Python utilities for handling LiPD data"
 readme = "README.md"
 license = "Apache-2.0"
