### 2025-06-29 95707d3
Updated paper.md
