### 2025-06-29 0ed506f
Bump version to v1.4.4
- Update CITATION.cff cff-version to 1.4.3.1 (newer Citation File Format specification)
- Add paper.bib with comprehensive bibliography for JOSS paper submission
- Complete academic publication package with paper.md and references
- Prepare PyLiPD for formal academic publication
