### 2025-06-30 bb6adda
Fix PyPI publishing - correct license metadata format
- Fix pyproject.toml license field format (license-files -> license = {text = 'Apache-2.0'})
- Remove invalid license-files field that was causing PyPI publishing to fail
- Bump version to v1.4.5 to trigger new release
- Update CITATION.cff to match new version

Fixes GitHub Action error: 'InvalidDistribution: unrecognized or malformed field license-file'
