+++ CITATION.cff
@@ -1,4 +1,4 @@
-cff-version: 1.4.4.1
+cff-version: 1.4.5.1
 message: "If you use this software, please cite it as below."
 authors:
   - family-names: "Ratnakar"
@@ -8,7 +8,7 @@ authors:
     given-names: "Deborah"
     orcid: "https://orcid.org/0000-0001-7501-8430"
 title: "PyLiPD: a Python package for the manipulation of LiPD datasets"
-version: v1.4.4
+version: v1.4.5
 doi: 10.5281/zenodo.7951201
 date-released: 2025-01-28
 url: "https://github.com/LinkedEarth/pylipd"
