### 2025-06-30 b2257dd
Bump version to v1.4.5
- pyproject.toml version to 1.4.5
- CITATION.cff version to v1.4.5 and cff-version 1.4.5.1
