+++ pyproject.toml
@@ -3,12 +3,13 @@ name = "pylipd"
 version = "1.4.5"
 description = "Python utilities for handling LiPD data"
 readme = "README.md"
-license = {text = "Apache-2.0"}
+license = "Apache-2.0"
 authors = [
   { name = "Varun Ratnakar", email = "varunratnakar@gmail.com" },
   { name = "Deborah Khider" }
 ]
 requires-python = ">=3.11"
+license-files = ["LICENSE"]
 dependencies = [
   "rdflib",
   "pandas",
