+++ .github/workflows/publish.yaml
@@ -1,9 +1,16 @@
 name: Upload Python Package
 
 on:
+  # Run automatically *once* when a release is published (not on draft/created/edited events)
   release:
-    branches: [ main ]
+    types: [published]
+  # Still allow manual trigger from the Actions tab
   workflow_dispatch:
+
+# Prevent duplicate runs for the same ref (e.g., if a release is republished)
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
 
 jobs:
   deploy:
@@ -15,7 +22,10 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install build twine
+        # Install latest packaging tooling so that PEP639 (license-expression / license-file)
+        # metadata fields are recognised by wheel and Twine.
+        pip install --upgrade build wheel
+        pip install --upgrade "twine>=5.0.0" "pkginfo>=1.9.7"
     - name: Build and publish
       env:
         TWINE_USERNAME: __token__
