### 2025-06-30 356a950
Remove license-files field from [project] (Twine metadata compatibility)
