+++ pyproject.toml
@@ -9,7 +9,6 @@ authors = [
   { name = "Deborah Khider" }
 ]
 requires-python = ">=3.11"
-license-files = ["LICENSE"]
 dependencies = [
   "rdflib",
   "pandas",
