+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pylipd"
-version = "1.4.5"
+version = "1.5.0"
 description = "Python utilities for handling LiPD data"
 readme = "README.md"
 license = { text = "Apache-2.0" }
