### 2025-06-30 aafb5d4
Packaging: suppress license-file header (PyPI compatibility)
