+++ scripts/analyze_compilations.py
@@ -4,25 +4,20 @@ Compilation Analysis Script for LiPDVerse
 
 This script:
 1. Queries GraphDB repository to get compilation names and dataset counts
-2. Downloads LiPD files for each compilation from LiPDverse
+2. Analyzes LiPD files from a local input directory  
 3. Loads LiPD files using PyLiPD and analyzes compilation assignments
 4. Generates reports on compilation consistency and coverage
 
 Usage:
-    python analyze_compilations.py [--output-dir OUTPUT_DIR] [--sparql-endpoint ENDPOINT]
+    python analyze_compilations.py --input-dir INPUT_DIR [--output-dir OUTPUT_DIR] [--sparql-endpoint ENDPOINT]
 """
 
 import os
 import sys
 import argparse
 import requests
-import zipfile
-import tempfile
 import json
 from pathlib import Path
-from urllib.parse import urljoin, urlparse
-import re
-from bs4 import BeautifulSoup
 import pandas as pd
 from collections import defaultdict, Counter
 import csv
@@ -31,14 +26,21 @@ from pylipd.lipd import LiPD
 
 
 class CompilationAnalyzer:
-    def __init__(self, sparql_endpoint, output_dir):
+    def __init__(self, sparql_endpoint, input_dir, output_dir):
         self.sparql_endpoint = sparql_endpoint
+        self.input_dir = Path(input_dir)
         self.output_dir = Path(output_dir)
         self.output_dir.mkdir(parents=True, exist_ok=True)
         
+        # Validate input directory
+        if not self.input_dir.exists():
+            raise ValueError(f"Input directory does not exist: {input_dir}")
+        if not self.input_dir.is_dir():
+            raise ValueError(f"Input path is not a directory: {input_dir}")
+        
         # Results storage
         self.graphdb_compilations = {}
-        self.downloaded_compilations = {}
+        self.local_compilations = {}
         self.lipd_analysis = {}
         
         # Mapping of dataset -> collections info
@@ -155,155 +157,46 @@ class CompilationAnalyzer:
             print(f"Error fetching dataset-compilation mapping: {e}")
             # Leave mapping empty if fails
 
-    def find_download_link(self, compilation_name):
+    def scan_input_directory(self):
         """
-        Find the download link for a specific compilation from LiPDverse.
-        Uses a multi-step search process:
-        1. First check the main page: https://lipdverse.org/[compilation]/current_version/
-        2. If not found, check the sidebar: https://lipdverse.org/[compilation]/current_version/projectSidebar.html
+        Scan the input directory for LiPD files and organize them by compilation.
         
-        Parameters
-        ----------
-        compilation_name : str
-            Name of the compilation
-            
         Returns
         -------
-        str or None
-            Download URL if found, None otherwise
+        dict
+            Dictionary mapping compilation names to directories containing LiPD files
         """
-        base_url = f"https://lipdverse.org/{compilation_name}/current_version/"
-        main_url = base_url
-        sidebar_url = f"{base_url}projectSidebar.html"
+        print(f"Scanning input directory: {self.input_dir}")
         
-        def search_for_download_links(url, soup):
-            """Helper function to search for download links in a BeautifulSoup object"""
-            # Look for "Download all LiPD files" link
-            for link in soup.find_all('a', href=True):
-                link_text = link.get_text(strip=True).lower()
-                if 'download all lipd files' in link_text or 'download all' in link_text:
-                    href = link['href']
-                    if href.startswith('http'):
-                        return href
-                    else:
-                        return urljoin(base_url, href)
-            
-            # Fallback: look for .zip files
-            for link in soup.find_all('a', href=True):
-                href = link['href']
[…]
