### 2025-07-02 fd5048e
Updated compilations download script
