+++ scripts/download_compilations.py
@@ -35,84 +35,158 @@ class LiPDCompilationDownloader:
         }
 
     def get_available_compilations(self):
-        """Get list of known compilations"""
-        return {
-            "CoralHydro2k",
-            "HoloCoral", 
-            "HoloceneAbruptChange",
-            "HoloceneHydroclimate",
-            "Hydro21k",
-            "LakeStatus21k",
-            "Pages2kTemperature",
-            "RapidArcticWarming",
-            "SISAL-LiPD",
-            "Temp12k",
-            "Temp24k",
-            "iso2k",
-            "wNAm"
-        }
+        """Get list of available compilations by parsing the LiPDverse project page"""
+        project_page_url = "https://lipdverse.org/project/"
+        
+        print("Fetching available compilations from LiPDverse...")
+        
+        try:
+            response = requests.get(project_page_url, timeout=30)
+            response.raise_for_status()
+            
+            soup = BeautifulSoup(response.content, 'html.parser')
+            
+            # Find compilation links following the pattern:
+            # <h2><a href="https://lipdverse.org/project/rapidarcticwarming/">RapidArcticWarming</a></h2>
+            compilations = {}
+            
+            for h2_tag in soup.find_all('h2'):
+                link_tag = h2_tag.find('a', href=True)
+                if link_tag and link_tag['href'].startswith('https://lipdverse.org/project/'):
+                    compilation_name = link_tag.get_text(strip=True)
+                    project_url = link_tag['href']
+                    
+                    # Skip generic links that aren't compilations
+                    if compilation_name and compilation_name not in ['Projects', 'Home', 'Articles']:
+                        compilations[compilation_name] = project_url
+            
+            print(f"Found {len(compilations)} available compilations:")
+            for name in sorted(compilations.keys()):
+                print(f"  - {name}")
+            
+            return compilations
+            
+        except Exception as e:
+            print(f"Error fetching compilations from project page: {e}")
+            print("Falling back to known compilations...")
+            
+            # Fallback to known compilations if the web request fails
+            known_compilations = {
+                "CoralHydro2k": "https://lipdverse.org/project/coralhydro2k/",
+                "HoloCoral": "https://lipdverse.org/project/holocoral/", 
+                "HoloceneAbruptChange": "https://lipdverse.org/project/holoceneabruptchange/",
+                "HoloceneHydroclimate": "https://lipdverse.org/project/holocenehydroclimate/",
+                "Hydro21k": "https://lipdverse.org/project/hydro21k/",
+                "LakeStatus21k": "https://lipdverse.org/project/lakestatus21k/",
+                "Pages2kTemperature": "https://lipdverse.org/project/pages2ktemperature/",
+                "RapidArcticWarming": "https://lipdverse.org/project/rapidarcticwarming/",
+                "SISAL-LiPD": "https://lipdverse.org/project/sisal/",
+                "Temp12k": "https://lipdverse.org/project/temperature12k/",
+                "Temp24k": "https://lipdverse.org/project/temp24k/",
+                "iso2k": "https://lipdverse.org/project/iso2k/",
+                "wNAm": "https://lipdverse.org/project/wnam/"
+            }
+            return known_compilations
 
-    def find_download_link(self, compilation_name):
+    def find_download_link(self, compilation_name, project_url):
         """
-        Find the download link for a specific compilation from LiPDverse.
+        Find the download link for a specific compilation using its project page URL.
         """
-        base_url = f"https://lipdverse.org/{compilation_name}/current_version/"
-        main_url = base_url
-        sidebar_url = f"{base_url}projectSidebar.html"
-        
-        def search_for_download_links(url, soup):
-            """Helper function to search for download links in a BeautifulSoup object"""
-            # Look for "Download all LiPD files" link
[…]
