### 2025-07-03 f7da61f
Release v1.5.1: Bug fixes and improvements
- Fix uniqid() function to use UUID4 for truly unique IDs
- Fix sorting bug in analyze_compilations.py when handling None values
- Add support for multiple versions in Compilation class
- Update schema to support multiple compilation versions
- Update scripts documentation and workflow instructions
