### 2025-07-03 6a8e149
Update CITATION.cff for v1.5.1 release
