### 2025-08-13 #8125 @willgearty
@PennyHow I've been unable to find another reviewer. Would you be available to review this JOSS submission?
