### 2025-08-17 7635b1c
Fix some typos in the documentation and code
I ended up running codepsell on the relevant files to try and be a
bit more systematic.
