### 2025-08-17 e850b06
Silence SyntaxWarnings in tests
