### 2025-08-17 3aa89a4
Streamline creation of development environment
