+++ pyproject.toml
@@ -22,6 +22,19 @@ dependencies = [
   "requests"
 ]
 
+[dependency-groups]
+dev = [
+  "IPython",
+  "jupyter-sphinx",
+  "nbsphinx",
+  "numpydoc>=1.1.0",
+  "readthedocs-sphinx-search>=0.1.0",
+  "pytest",
+  "Sphinx",
+  "sphinx-copybutton",
+  "sphinx-rtd-theme>=1.0.0"
+]
+
 [tool.setuptools.packages.find]
 where = ["."]
 include = ["pylipd*"]
