### 2025-08-25 #8125 @PennyHow
> [@PennyHow](https://github.com/PennyHow) I've been unable to find another reviewer. Would you be available to review this JOSS submission?

Hi @willgearty, sorry for the slow response - I've been on holiday. I'm back now, and I am available to review this. I also have another JOSS review but it is still at the pre-review stage; so I think it is fine to do this one as well 👍🏼
