### 2025-08-28 LinkedEarth/pylipd#83 @spencerkclark
Tutorials page undersells the `pylipdTutorials` repository
When one clicks on the link to the tutorials in the documentation, one is taken to a page that blandly states:

> Tutorials for PyLiPD exist in the form of:
> 	- Jupyter notebooks, which are found on the following repositories:
> 		- [https://github.com/LinkedEarth/pylipdTutorials](https://github.com/LinkedEarth/pylipdTutorials)

As a person arriving at PyLiPD and the LiPD ecosystem for the first time, I found this tutorial incredibly helpful.  This is more than just a collection of Jupyter notebooks in a GitHub repository.  It is organized and published via JupyterBook, and includes a high-level description of the LiPD Ontology, what the Resource Description Framework and SPARQL query language are, and some background motivation for the overall project, on top of some very nice usage examples.

At a minimum I think a link directly to the JupyterBook should be present on the landing page of the documentation.  If possible it would also be helpful to include a prominent link directly to the JupyterBook in the sidebar as well.  Unless there are plans to eventually write more tutorial material at separate locations, one could consider removing the dedicated "Tutorials" page in the documentation, which currently makes it feel somewhat like an afterthought (when it clearly is not!).

xref: https://github.com/openjournals/joss-reviews/issues/8861
