### 2025-08-28 LinkedEarth/pylipd#84 @spencerkclark
Mention deprecation of previous `LiPD-utilities` package
When browsing various websites to learn about the LiPD framework, I came across the [`LiPD-utilities` repository](https://github.com/nickmckay/LiPD-utilities), which makes a vague reference in the README to the fact that:

> The python package is also in the process of being replaced with a more modern architecture.

It would be helpful to close the loop here and mention that this is that new package, for example by mentioning the old `LiPD-utilities` package on the documentation landing page as well as potentially in the README, perhaps also including some background as to why a new package was needed.  Likewise it would be useful to update the README of the `LiPD-utilities` package with a link to this repository.  That way if anyone inadvertently comes across it, they will know that is an old package and know to come here instead.

xref: https://github.com/openjournals/joss-reviews/issues/8861
