### 2025-08-29 5e50473
readthedocs badge added
