### 2025-08-29 c2fa255
bullet point formatting altered for correct readthedocs rendering
