+++ docs/tutorials.rst
@@ -3,10 +3,157 @@
 Tutorials
 =========
 
-Tutorials for PyLiPD exist in the form of:
+This page contains a few tutorials for using PyLiPD.  
 
-* Jupyter Notebooks, which are found on the following repositories:
+.. grid:: 3
+   :gutter: 2
+   :margin: 2
 
-  * https://github.com/LinkedEarth/pylipdTutorials
+   .. card::
+      :img-top: _static/thumbs/01_intro.png
+      :link: https://linked.earth/pylipdTutorials/intro.html
+      :text-align: left
+
+      **An Introduction to the PyLiPD Package**
+      Provides an overview of the package and the tutorials
+
+
+Dataset Representation in PyLiPD
+""""""""""""""""""""""""""""""""
+
+We recommend reading these tutorials if you want to get a deeper understanding of how the datasets are represented. These tutorials are mandatory reading for anyone creating custom queries or needing to create/edit a LiPD file.
+
+.. grid:: 3
+   :gutter: 2
+   :margin: 2
+
+   .. card::
+      :img-top: _static/thumbs/02_DataRep.png
+      :link: https://linked.earth/pylipdTutorials/graph.html
+      :text-align: left
+
+      **Understanding how datasets are represented in PyLiPD**
+      This tutorial provides an overview of the ontology and the SPARQL language
+
+   .. card::
+      :img-top: _static/thumbs/03_Stds.png
+      :link: https://linked.earth/pylipdTutorials/standards.html
+      :text-align: left
+
+      **Standard, Standards, and More Standards**
+      This tutorial provides an overview of the vocabulary used for the values of certain terms such as archive and proxy.
+      
+Getting Started
+"""""""""""""""
+.. grid:: 3
+   :gutter: 2
+   :margin: 2
+
+   .. card::
+      :img-top: _static/thumbs/04_Loading.png
+      :link: https://linked.earth/pylipdTutorials/notebooks/L0_a_loading_lipd_datasets.html
+      :text-align: left
+
+      **Reading LiPD formatted datasets with PyLiPD**
+      This tutorial provides examples on how to load local files, files in a directory, from a URL, and from the GraphDB. 
+
+   .. card::
+      :img-top: _static/thumbs/05_Manip.png
+      :link: https://linked.earth/pylipdTutorials/notebooks/L0_b_lipd_object.html
+      :text-align: left
+
+      **Basic Manipulation of pylipd.LiPD objects**
+      This tutorial shows how to extract information (data and associated metadata) from LiPD files and how to add and remove datasets from an existing object.
+
+
+Basic Functionalities
+"""""""""""""""""""""
+.. grid:: 3
+   :gutter: 2
+   :margin: 2
+
+   .. card::
+      :img-top: _static/thumbs/06_Ensembles.png
+      :link: https://linked.earth/pylipdTutorials/notebooks/L1_a_working_with_ensembles.html
+      :text-align: left
+
+      **Working with ensembles in PyLiPD**
+      This tutorial describes how PyLiPD deals with age ensembles often used for uncertainty quantification in paleoclimate studies. 
+
+   .. card::
+      :img-top: _static/thumbs/07_Text.png
+      :link: https://linked.earth/pylipdTutorials/notebooks/L1_b_getting_information.html
+      :text-align: left
+
+      **Retrieving textual information from LiPD files**
+      This tutorial demonstrates the use of pre-defined APIs to get specific information from a LiPD file. 
+   
+   .. card::
+      :img-top: _static/thumbs/08_Filters.png
+      :link: https://linked.earth/pylipdTutorials/notebooks/L1_c_filtering.html
+      :text-align: left
+
+      **Filtering through queries**
[…]
