+++ docs/rtd_env.yml
@@ -18,4 +18,5 @@ dependencies:
     - sphinx-rtd-theme>=1.0.0
     - jupyter-sphinx
     - sphinx-copybutton
+    - sphinx-design
     - git+https://github.com/LinkedEarth/pylipd@main

[…]
