### 2025-10-06 6671bee
Paper edits
