### 2025-10-06 #8861 @khider
Hi @PennyHow, sorry for the delay on this.

I added thumbnails links to the tutorials to the documentation: https://pylipd.readthedocs.io/en/latest/tutorials.html  so they feel more integrated. We expect to have tutorials that may be too heavy to build on readthedocs, hence the reason we chose to link them instead.

Other comments:

### Functionality

We changed the documentation to reflect the fact that PyLiPD is compatible with version 3.11 and above

### Documentation

1. To avoid rebuilding the notebooks inside readthedocs containers, use thumbnails tutorials to the GitHub repo: https://pylipd.readthedocs.io/en/latest/tutorials.html
2. The usage section has been added to the installation documentation: https://pylipd.readthedocs.io/en/latest/installation.html#usage
3.
	1. Dependencies added: https://pylipd.readthedocs.io/en/latest/installation.html#dependencies
	2. Instructions on how to add execute the pytest tests: https://pylipd.readthedocs.io/en/latest/installation.html#running-the-test-suite
	3. Creating an environment: https://pylipd.readthedocs.io/en/latest/installation.html#creating-an-environment
	4. Developer instructions (from GitHub install): https://pylipd.readthedocs.io/en/latest/installation.html#installing-pylipdl with a link to the contributing guide, which further details how to contribute to the package: https://pylipd.readthedocs.io/en/latest/contribution_guide.html
4. Pull request was accepted and changes merged into the main branch.

### Software paper

1. Added a citation for the R tool to manipulate LiPD files. Pyleoclim and cfr are added to the statement of need as possible libraries to use with these datasets.
2. Changed
3. Added (Note that SPARQL is a recursive acronym so the S stands for SPARQL)
4. There was a typo in the full citation black (used ',' instead of ";" to separate the various studies)
5. The modules and names have been clarified.
