### 2025-10-16 5548b73
Update CITATION.cff for version 1.5.2
