+++ CITATION.cff
@@ -1,4 +1,4 @@
-cff-version: 1.5.1.1
+cff-version: 1.5.2.1
 message: "If you use this software, please cite it as below."
 authors:
   - family-names: "Ratnakar"
@@ -8,7 +8,8 @@ authors:
     given-names: "Deborah"
     orcid: "https://orcid.org/0000-0001-7501-8430"
 title: "PyLiPD: a Python package for the manipulation of LiPD datasets"
-version: v1.5.1
+version: v1.5.2
 doi: 10.5281/zenodo.7951201
-date-released: 2025-07-03
+date-released: 2025-10-16
+license: Apache-2.0
 url: "https://github.com/LinkedEarth/pylipd"
