### 2025-10-16 5a16c55
Bump version from 1.5.1 to 1.5.2
