### 2025-10-24 f324be0
Update paper.md
