+++ paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 `PyLiPD` is a Python package designed to support the reading, querying, editing, and writing of paleoclimate datasets formatted in the Linked Paleo Data Format (LiPD) [@McKay2016]. Built on the `rdflib` library, it transforms LiPD data into Resource Description Framework (RDF) graphs aligned with the LinkedEarth ontology, enabling semantic querying with the SPARQL Protocol and RDF Query  (SPARQL). The API is structured into four modules that handle data access, variable-level filtering, and dataset creation and editing, with support for both graph-based and tabular workflows. Comprehensive documentation and tutorials—including ontology concepts and scientific use cases—are provided to guide users. 
 
 # Statement of Need
-Paleoclimate data (obtained from biogeophysical measurements made on ice cores, sediments, corals, trees, etc.) often come with diverse structures and metadata conventions, making synthesis efforts challenging and highlighting the necessity of standardized data formats.  Achieving standardization involves three key components: (1) a uniform data format, (2) a consistent terminology for describing metadata, and (3) clear guidelines for how data should be reported. The LiPD framework emerged to address the first need, providing a universally readable data container for paleoclimate data and metadata [@McKay2016]. The metadata is stored in a JSON-LD file while the data is organized in multiple tables saved as csv files. LiPD has six distinct components: root metadata (e.g., dataset name, and version); geographic metadata (e.g., coordinates); publication metadata (including unique identifiers); funding metadata (e.g., grant number); PaleoData, which includes all the measured (e.g., the width of tree rings) and inferred (e.g., temperature) paleoenvironmental data; and ChronData, which mirrors PaleoData for information pertaining to time. These components provide the rigidity necessary to write robust codes around the format while remaining extensible enough to capture (meta)data as rich as the users want to provide for them. Paired with the Paleoenvironmental Standard Terms (PaST) Thesaurus developed by the National Oceanic and Atmospheric Administration (NOAA) [@Morrill2021], LiPD standardization facilitates efficient querying and analysis of multiple records, enabling large-scale paleoclimate syntheses. Community-driven initiatives to reconstruct temperature and hydroclimate over Earth's history [@pages2k; @Kaufman2020; @Routson2021; @Konecky2020; @Jonkers2020] have compiled paleoclimate records into this standardized format and have made them available to the paleoclimate community in an online database. As the database continued to grow, there was an increasing demand for tools capable of accessing, reading, modifying, and writing files in the LiPD format. Although there is a package in R [@LipdR], a Python tool built for scientists and interoperable with libraries such as `Pyleoclim` [@Khider2022] for time series analysis and `cfr` [@Zhu2024] for climate field reconstruction is lacking. To address this need, we introduce `PyLiPD`, a Python-based tool designed specifically for these tasks.
[…]
