### 2025-10-27 #8861 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
