### 2025-10-29 #8861 @kthyng
@khider Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "earth" and "python" among others.
